requirements.txt
Flask==2.0.1
SQLAlchemy==1.4.36
Flask-SQLAlchemy==2.5.1
Flask-Mail==0.9.1
itsdangerous==2.0.1
Werkzeug==2.0.1
python-dotenv==0.19.0